# docs/swagger.yaml
basePath: /
definitions:
  types.ProductDTO:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
  types.UserDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: Product management api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Api to create products
  version: "1.0"
paths:
  /api/v1/products:
    delete:
      consumes:
      - application/json
      description: Receives an id and delete a product by given id
      parameters:
      - description: access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: "1"
        format: uuid
        in: query
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete product by id
      tags:
      - products
    get:
      consumes:
      - application/json
      description: List all the products paginated
      parameters:
      - description: access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: "1"
        in: query
        name: page
        required: true
        type: integer
      - description: "10"
        in: query
        name: size
        required: true
        type: integer
      - description: asc|desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Find all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Endpoint to create a product
      parameters:
      - description: access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: product.request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.ProductDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Endpoint to update a product
      parameters:
      - description: access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: product.request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.ProductDTO'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update product
      tags:
      - products
  /api/v1/products/get:
    get:
      consumes:
      - application/json
      description: Receives an id and returns a product by given id
      parameters:
      - description: access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: "1"
        format: uuid
        in: query
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get product by id
      tags:
      - products
  /login:
    post:
      consumes:
      - multipart/form-data
      description: Receives an email and a password and returns a JWT
      parameters:
      - description: email
        in: formData
        name: email
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Endpoint to authenticate
      tags:
      - users
  /sign-up:
    post:
      consumes:
      - application/json
      description: Receives a request and creates a new user
      parameters:
      - description: users.request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.UserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a new user
      tags:
      - users
swagger: "2.0"